Move `getIgnoredReport` to `report.js`
