+++ lib/report.js
@@ -67,4 +67,18 @@ const getReportStatistics = results => {
 	return statistics;
 };
 
-export {mergeReports, processReport};
+const getIgnoredReport = filePath => ({
+	errorCount: 0,
+	warningCount: 0,
+	results: [
+		{
+			errorCount: 0,
+			warningCount: 0,
+			filePath,
+			messages: [],
+		},
+	],
+	isIgnored: true,
+});
+
+export {mergeReports, processReport, getIgnoredReport};
